Refactored HealthSummary related components to allow usage in home and in components overviews
