+++ assets/js/components/HealthSummary/ComponentHealthSummary.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import HealthSummary from './HealthSummary';
+
+export const ComponentHealthSummary = ({ data }) => {
+  return (
+    <div className="mb-8">
+      <HealthSummary data={data} />
+    </div>
+  );
+};
